Use sync_utils to re-write received tar archive
What makes it different from the previous iteration is that this file,
since it's no longer a simple copy, is also processed to extract the
manifest and is archived.

The manifest is by no means the exact direction we'll be taking, but
rather a placeholder for the time being.

I've also extracted stream->sync functionality into
`ByteStreamReceiverAsyncSender`

The API is not perfect, but it makes code much easier to read.