Problem: trunk install doesn't install anything
Solution: install archives made by `trunk build`

If it's a tar, it'll work directly over those, if it is `tar.gz`,
it will unpack those into a tempfile first.

The reason for this is that we always want to work with tar a file
because `File` implements `Seek` and allows us to process the archives
more efficiently, without reading what we don't need in full.

The extraction happens in two passes. In the first pass we find
the manifest. In the second, we extract files declared in the manifest.

I've changed the manifest format to list files in an object as opposed
to an array. This simplifies its processing, and also is slightly more
correct as it won't allow for duplicate file names. That being said, the
format of manifest is still very much work-in-progress.