Update checkout and rust toolchain actions
Replace actions/checkout@v3 with @v4; the former is deprecated and
issues warnings for an out-of-date version of Node JS.

Replace actions-rs/toolchain with dtolnay/rust-toolchain, as the former
has had no changes in 4 years, uses the deprecated `set-output`
function, and was archived in October. Follow the findings of
RustCrypto/actions#17.