This is the inital "fix" for plat-467. The primary modifications come from following the documentation for the CNPG declarative hibernation feature:
https://cloudnative-pg.io/documentation/current/declarative_hibernation/

This introduces the cnpg.io/hibernation annotation and sets it to "on" when
the spec.stop attribute is true. Currently this only circumvents reconcile
blocks which depend on the cluster to be running in some way, but I suspect
it would be possible to be more aggressive depending on how much extra
operator overhead we want to remove. This reactivates the "deprecated" code
surrounding the spec.stop attribute; former attempts to use this flag resulted
in a cluster permanently stuck in a warning state.