Added the missing code which should allow us to thaw a hibernated cluster. Also fixed a possible bug where the pods_to_fence variable masks a function of the same name.
Unfortunately this will not work as written. Once we unset the "stop" flag,
all of our existing circumvention will fail to restart the cluster because
of the current fencing code. The previous spec.stop state will be lost, and
the operator will immediately try to fence pods that don't exist. This means
the fencing code needs to check for the hibernation flag. This code will be
added later.