Re-synced branch with main.
Fixed the issue where the cluster can't thaw after spec.stop is
set to false. The issue was that pods_to_fence is designed to be
called before cnpg_cluster_from_cdb where we toggle the hibernation
annotation. This function implicitly seeks a Primary pod, but since
there are no pods, it simply enters a permanent retry loop while
expecting the Primary pod to appear. Now we check directly for the
hibernation annotation here and in reconcile_cnpg to guarantee a
code path to the annotation.