Fixed the canonical name of the hibernate test.
Converted the basic-pgsql test to use the TestCore base struct. A focused invocation shows this test is still functional post conversion.

Removed state field from TestCore struct, as it's just a means of obtaining an Arc context.

Further investigation shows the "pods" attribute almost never comes without a named pod instance, but it is used in isolation in nested calls. Later refactors should account for this and add another field to the struct for this named pod instance (usually named {cluster-name}-1). The TestCore class is serving primarily as an implementation example for now.