Reduce test thread concurrency
The tests were being run with this command:

```
cargo test --jobs 1 -- --ignored --nocapture
```

However it should be noted that `--jobs 1` only affects compilation, not the amount of concurrency. We were likely overloading the test system and causing timeouts. We could probably use 4-8 threads, but I'm setting this to 1 since that seems to have been the initial intent.