Add support for CNPG Hibernation functionality (#720)
This patch set should make it possible to pause an entire CoreDB cluster and all related components such as app services. This works by setting the spec.stop setting to a boolean true value, and will result in the CNPG operator being put in hibernation mode using the cnpg.io/hibernation annotation, and all app service deployments set to 0 replicas. These changes are reverted by setting spec.stop to false. This should make it possible to eventually add an automated system to pause an otherwise inactive cluster, or to do so on demand.

Co-authored-by: Nick Hudson <nick.hudson@gmail.com>
Co-authored-by: Steven Miller <sjmiller609@gmail.com>