Remove OLAP, DataWarehouse (#1123)
Remove the stacks and all references to them. Also remove remaining
references to the Rag stack, removed in 6d472c9 (#1070).

While at it, update the integration test environment variable
configuration to default to using the dev API and testing org ID (also
present in `cargo-test.yml`) and returning a more informative error when
the `ACCESS_TOKEN` isn't set. Also fix the lifetime scoping of the
return value of `gradient_p()` and delete a test assertion that the
compiler removes.