Upgrade all PostgREST images to 12.2.8 (#1150)
For comments, examples, and docs, use the `:latest` tags. Everywhere
else, explicitly use `:v12.2.8`.

Except in `tembo-operator`, where the tests just will not pass. But
disable Grafana in the operator tests, it's not necessary.

Also fix the misspelling of `Authorization` in a test (and let the
linter reformat the relevant code).