Remove pg_tier and parquet_s3_fdw from stacks (#1160)
`pg_tier` does not work (tem-2870), so remove it. Also remove
`parquet_s3_fdw`, since it was only included to support `pg_tier`, and
its functionality is covered by `pg_duckdb`, `pg_analytics`, and
`duckdb_fdw`.

Also switch the analytics stack to the `standard-cnpg` image. The only
thing different about the `analytics-cnpg` image is that it included
Tembo's for of `pg_analytics`, which is no longer needed.