Refactor shell commands on container init (#)
Move the shell commands to a script, `init-dirs.sh`, and embed it in the
binary. This makes it easier to maintain the shell script without Rust
escapes and no line breaks.

Change the script behavior to not only copy the contents of
`/tmp/pg_sharedir` if sharedir is empty, but also always copy the
contents of `/tmp/pg_pkglibdir`, as a minor version upgrade might
include improvements or bug fixes.