Eliminate need for `/tmp/pg_sharedir` (#1185)
Mount the `pgdata` volume in the `tembo-init` container to
`/var/lib/postgresql/init` instead of `/var/lib/postgresql/data`. This
leaves the contents of `data` from the image in place, and we can simply
copy them to `init`, eliminating the need for the images to have copied
data to `/tmp/pg_sharedir`.

Update `init-dirs.sh` to make this copy, but switch to `cp --recursive
--update`, which will update files but not delete any other files. This
will allow change such as minor Postgres version updates to shared
libraries to be copied over, ensuring they're consistent with the
version of Postgres from the image.

Remove the `/tmp/pg_pkglibdir` copy, as well, as it is included in
`data`, and simplify the deletion of
`/var/lib/postgresql/data/lost+found` (which should not be in any of our
images anyway).