Add pull_policy to pod init Config (#1200)
Add the `uses_postgres_image()` method to `Config` in the
`tembo-pod-init` package. Like the method with the same name in
`CoreDBSpec` in `tembo-operator`, the method returns true when the image
is a base image and false when it's not.

Also add `image_pull_policy()`, which returns returns "Always" when the
image name is `postgres` and its tag includes a Postgres major version
and OS name. This ensures that on every pod init, it will pull the
latest image for that OS and Postgres version. For all other cases it
only pulls "IfNotExists".

Increment tembo-pod-stacks to v0.2.1.