Install all extensions on Postgres images (#1201)
Modify `find_trunk_installs_to_pod` to return the full list of Trunk
packages to install when the image is the new Postgres base image. This
ensures that any rebuilds of the extension will be applied, as well as
any dependencies (once #1199) is committed. The goal is to always have
the latest builds when a pod starts, so to keep up-to-date with any
upstream security fixes.